% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChocoRatingPackage-package.R
\docType{package}
\name{ChocoRatingPackage}
\alias{ChocoRatingPackage-package}
\alias{ChocoRatingPackage}
\title{ChocoRatingPackage: Tools for Exploring Chocolate Ratings with Shiny App}
\description{
The \code{ChocoRatingPackage} provides chocolate ratings dataset and an interactive Shiny app to explore chocolate ratings
by company and country of origin. Users can filter and visualize chocolate ratings data, gaining
insights into chocolate quality across different manufacturers and regions.
}
\section{Data}{

\itemize{
\item \code{chocolate_rating}: A dataset containing chocolate ratings information.
}
}

\section{Function}{

\itemize{
\item \code{explore_ratings()}: Launches the Shiny app for exploring chocolate ratings.
}
}

\author{
\strong{Maintainer}: Anh Phan \email{hpha0054@student.monash.edu}

}
