% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{explore_ratings}
\alias{explore_ratings}
\title{Launch the Chocolate Rating Shiny App}
\usage{
explore_ratings()
}
\value{
This function does not return a value. It launches a Shiny application
that opens in a new window or browser tab.
}
\description{
The \code{explore_ratings()} function launches an interactive Shiny application that
allows users to explore chocolate ratings by company and country of origin.
This app displays a boxplot of chocolate ratings based on a user-selected country,
with additional controls to filter companies by the minimum number of ratings.
It provides an intuitive interface for discovering statistics rating in chocolate quality
across different manufacturers and countries.
}
\details{
The Shiny app features:
\itemize{
\item An input field for specifying the country of origin.
\item A slider to adjust the minimum number of ratings required for a company
to be displayed.
\item Interactive hover-over tooltips on the boxplot to show detailed rating
information for each company.
}
}
